--- bricks/postgrest_errors/__brick__/HELLO.md
@@ -1,1 +1,0 @@
-Hello {{name}}!
