+++ bricks/postgrest_errors/hooks/pre_gen.dart
@@ -1,0 +1,3 @@
+import 'package:mason/mason.dart';
+
+void run(HookContext context) {}
