+++ bricks/postgrest_errors/hooks/pubspec.yaml
@@ -1,0 +1,8 @@
+name: hooks
+publish_to: none
+
+environment:
+  sdk: ">=3.1.0 <4.0.0"
+
+dependencies:
+  mason: ">=0.1.0-dev.50 <0.1.0"
