--- packages/postgrest_errors_scraper/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:very_good_analysis/analysis_options.5.1.0.yaml
