+++ packages/postgrest_errors_scraper/pubspec.yaml
@@ -11,6 +11,7 @@ dependencies:
   html: ^0.15.4
   http: ^1.1.0
   meta: ^1.11.0
+  path: ^1.8.3
 
 dev_dependencies:
   mocktail: ^1.0.0
